pyproject.toml
[tool.poetry]
name = "resume-system"
version = "0.1.0"
description = ""
authors = ["temirovazat <azatbekt5@gmail.com>"]
readme = "README.md"
packages = [{include = "resume-system"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
envparse = "^0.2.0"
drf-yasg = "^1.21.5"
django-filter = "^23.2"
black = "^23.3.0"
isort = "^5.12.0"
bandit = "^1.7.5"
flake8 = "^6.0.0"
django-cors-headers = "^3.14.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
